Rewrite block properties
