+++ pumpkin-world/src/block/mod.rs
@@ -7,7 +7,7 @@ use pumpkin_util::math::vector3::Vector3;
 pub use block_state::BlockState;
 
 #[derive(FromPrimitive, PartialEq, Clone, Copy)]
-pub enum BlockFace {
+pub enum BlockDirection {
     Bottom = 0,
     Top,
     North,
@@ -18,7 +18,7 @@ pub enum BlockFace {
 
 pub struct InvalidBlockFace;
 
-impl TryFrom<i32> for BlockFace {
+impl TryFrom<i32> for BlockDirection {
     type Error = InvalidBlockFace;
 
     fn try_from(value: i32) -> Result<Self, Self::Error> {
@@ -34,15 +34,15 @@ impl TryFrom<i32> for BlockFace {
     }
 }
 
-impl BlockFace {
+impl BlockDirection {
     pub fn to_offset(&self) -> Vector3<i32> {
         match self {
-            BlockFace::Bottom => (0, -1, 0),
-            BlockFace::Top => (0, 1, 0),
-            BlockFace::North => (0, 0, -1),
-            BlockFace::South => (0, 0, 1),
-            BlockFace::West => (-1, 0, 0),
-            BlockFace::East => (1, 0, 0),
+            BlockDirection::Bottom => (0, -1, 0),
+            BlockDirection::Top => (0, 1, 0),
+            BlockDirection::North => (0, 0, -1),
+            BlockDirection::South => (0, 0, 1),
+            BlockDirection::West => (-1, 0, 0),
+            BlockDirection::East => (1, 0, 0),
         }
         .into()
     }
