--- pumpkin/src/block/properties/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod slab;
-pub(crate) mod stair;
